更新README

+++ WeChatGUI/WeChatDLL/ApiServer.h
@@ -1,4 +1,6 @@
 #pragma once
 
 
-void StartApiServer(unsigned short port);
+void StartApiServer(int port);
+
+
